move stategen mock.go to its on testonly pkg
