+++ beacon-chain/rpc/prysm/v1alpha1/beacon/BUILD.bazel
@@ -97,6 +97,7 @@ go_test(
         "//beacon-chain/p2p/testing:go_default_library",
         "//beacon-chain/state:go_default_library",
         "//beacon-chain/state/stategen:go_default_library",
+        "//beacon-chain/state/stategen/mock:go_default_library",
         "//beacon-chain/state/v1:go_default_library",
         "//beacon-chain/sync/initial-sync/testing:go_default_library",
         "//cmd:go_default_library",
