+++ beacon-chain/rpc/prysm/v1alpha1/validator/BUILD.bazel
@@ -125,6 +125,7 @@ go_test(
         "//beacon-chain/powchain/testing:go_default_library",
         "//beacon-chain/state:go_default_library",
         "//beacon-chain/state/stategen:go_default_library",
+        "//beacon-chain/state/stategen/mock:go_default_library",
         "//beacon-chain/state/v1:go_default_library",
         "//beacon-chain/sync/initial-sync/testing:go_default_library",
         "//config/features:go_default_library",
