+++ beacon-chain/rpc/prysm/v1alpha1/validator/status_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	mockPOW "github.com/prysmaticlabs/prysm/beacon-chain/powchain/testing"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state"
-	"github.com/prysmaticlabs/prysm/beacon-chain/state/stategen"
+	mockstategen "github.com/prysmaticlabs/prysm/beacon-chain/state/stategen/mock"
 	v1 "github.com/prysmaticlabs/prysm/beacon-chain/state/v1"
 	mockSync "github.com/prysmaticlabs/prysm/beacon-chain/sync/initial-sync/testing"
 	fieldparams "github.com/prysmaticlabs/prysm/config/fieldparams"
@@ -937,7 +937,7 @@ func TestServer_CheckDoppelGanger(t *testing.T) {
 			name:    "normal doppelganger request",
 			wantErr: false,
 			svSetup: func(t *testing.T) (*Server, *ethpb.DoppelGangerRequest, *ethpb.DoppelGangerResponse) {
-				mockGen := stategen.NewMockService()
+				mockGen := mockstategen.NewMockService()
 				hs, ps, os, keys := createStateSetup(t, 4, mockGen)
 				// Previous Epoch State
 				for i := 0; i < 3; i++ {
@@ -982,7 +982,7 @@ func TestServer_CheckDoppelGanger(t *testing.T) {
 			name:    "doppelganger exists current epoch",
 			wantErr: false,
 			svSetup: func(t *testing.T) (*Server, *ethpb.DoppelGangerRequest, *ethpb.DoppelGangerResponse) {
-				mockGen := stategen.NewMockService()
+				mockGen := mockstategen.NewMockService()
 
 				hs, ps, os, keys := createStateSetup(t, 4, mockGen)
 				// Previous Epoch State
@@ -1049,7 +1049,7 @@ func TestServer_CheckDoppelGanger(t *testing.T) {
 			name:    "doppelganger exists previous epoch",
 			wantErr: false,
 			svSetup: func(t *testing.T) (*Server, *ethpb.DoppelGangerRequest, *ethpb.DoppelGangerResponse) {
-				mockGen := stategen.NewMockService()
+				mockGen := mockstategen.NewMockService()
 
 				hs, ps, os, keys := createStateSetup(t, 4, mockGen)
 				// Previous Epoch State
@@ -1116,7 +1116,7 @@ func TestServer_CheckDoppelGanger(t *testing.T) {
 			name:    "multiple doppelganger exists",
 			wantErr: false,
 			svSetup: func(t *testing.T) (*Server, *ethpb.DoppelGangerRequest, *ethpb.DoppelGangerResponse) {
-				mockGen := stategen.NewMockService()
+				mockGen := mockstategen.NewMockService()
 
 				hs, ps, os, keys := createStateSetup(t, 4, mockGen)
 				// Previous Epoch State
@@ -1166,7 +1166,7 @@ func TestServer_CheckDoppelGanger(t *testing.T) {
 			name:    "attesters are too recent",
 			wantErr: false,
 			svSetup: func(t *testing.T) (*Server, *ethpb.DoppelGangerRequest, *ethpb.DoppelGangerResponse) {
-				mockGen := stategen.NewMockService()
+				mockGen := mockstategen.NewMockService()
 
 				hs, _, _, keys := createStateSetup(t, 4, mockGen)
 
@@ -1213,7 +1213,7 @@ func TestServer_CheckDoppelGanger(t *testing.T) {
 	}
 }
 
-func createStateSetup(t *testing.T, head types.Epoch, mockgen *stategen.MockStateManager) (state.BeaconState,
+func createStateSetup(t *testing.T, head types.Epoch, mockgen *mockstategen.MockStateManager) (state.BeaconState,
 	state.BeaconState, state.BeaconState, []bls.SecretKey) {
 	gs, keys := util.DeterministicGenesisState(t, 64)
 	hs := gs.Copy()
