+++ beacon-chain/state/stategen/mock/mock.go
@@ -1,4 +1,4 @@
-package stategen
+package mock
 
 import (
 	"context"
